+++ x509/src/ext/pkix/constraints/mod.rs
@@ -2,4 +2,7 @@
 
 mod basic;
 
+pub mod name;
+
 pub use basic::BasicConstraints;
+pub use name::NameConstraints;
